The fixed PSF is given as double so we get full precision when calculating the PSF in python code
git-svn-id: http://svn.code.sf.net/p/bioimagexd/code/home/svnuser/repos/selli/trunk@658 b2d525f8-dbc2-4032-a3a1-b86af67c733b

+++ C++/vtkImageColocalizationTest.h
@@ -60,8 +60,8 @@ public:
 
   // Description:
   // Set the size of the PSF (in pixels) manually
-  vtkGetMacro(ManualPSFSize,int);
-  vtkSetMacro(ManualPSFSize,int);
+  vtkGetMacro(ManualPSFSize,double);
+  vtkSetMacro(ManualPSFSize,double);
 
   // Description:
   // Set the number of iterations
@@ -144,7 +144,7 @@ private:
   int IgnoreZeroPixels;
   int RandomizeZ;
   int Smooth;
-  int ManualPSFSize;
+  double ManualPSFSize;
   int NumIterations;
   int Ch2Lambda;
   double NumericalAperture;
